Create "feature manifest" feature spec
As we get closer to shipping this product, it's important to verify that
our features work as expected.

This is the first step to adding test coverage to the application. It
consists of test stubs for all of the existing "happy path" features.
Putting all of this in one file right now gives us a high-level view of
the application's features and lets us discuss what we're missing from
the list (/application).

As the test stubs get filled out, they should be moved into more focused
feature test files.