Add feature specs related to pledging items
In this feature, I left a number of specs unimplemented and avoided
filling out the `PledgesController` unit tests. Since this feature is so
early in its implementation, I don't want to lock us into the wrong
functionality. The other specs can be implemented as we decide how
pledging items is going to work.