Move API response logic out of controller
Previously, we were making the API call, parsing the XML response, and
sending the parsed hash to a response object all in the controller. That
logic isn't controller specific, and really belonged in `lib`.

This commit moves the response parsing logic from the controller action
into the AmazonProductAPI library. Now, the controller just needs to
know about the following method:

  ```
  AmazonProductAPI::HTTPClient.new(...).search_response
  #=> <AmazonProductAPI::SearchResponse object>
  ```

It's no longer concerned with the format of the API response!

Note:

  - I removed the Nokogiri::XML.parse call because I didn't think it was
    needed. If that was there as a bugfix, let me know!