Remove scaffolded API
Resolves #57

The Rails scaffold generator generates code for a json api in the
controller and the views. Because of this, we had an out-of-date
partially-implemented API cluttering up our controller.

This commit removes the `respond_to` blocks and the `.json.jbuilder`
files so that our controllers are easier to read and maintain.

If we choose to have an API, we can add it deliberately in the future.