Whitelist pages for non-admin users
Previously, we were explicitly marking admin pages with
`authenticate_admin`. This works, but if we forget to add the before
action to a secure controller action, we could share potentially
sensitive information with non-admins.

This commit changes this to a whitelist system–unless a
`skip_before_filter :authenticate_admin` is called, a page is
"admin-only" by default.