Add foreign keys and indices to relationships
Relationships were added to the database as integers. This works, but it
means we lose out on db-level referential integrity (keeping the records
in sync) and indexing on the model relationships. This commit adds these
back in.

For future migrations that deal with relations, I suggest the following
generator:

    ```bash
    $ bin/rails generate AddThingAToThingB thing_a:belongs_to
    ```

or the following for the migration:

    ```ruby
    def change
      add_reference :thing_a, :thing_b, foreign_key: true
    end
    ```