Use pundit for authorization
This change allows for more complex authorization logic without
cluttering up the controllers. It also allows for more isolated testing
of authorization policies.

Authorizing logged in users and admin aren't too bad in the controller,
but we're also checking for site managers for specific sites and users
owning their own profiles, which makes the controllers busy.