Use pry for rails console
We had `pry` in the Gemfile, but we were still using irb for the rails
console. This commit changes that to pry using the `pry-rails` wrapper.

Pry has a number of features missing in irb (such as navigation through
classes and methods, better autocomplete, and better syntax
highlighting) and is all-around more pleasant to use.

I chose `pry-rails` (over doing it by hand) because it's a very
lightweight, readable library, and it adds a few nice convenience
methods (ex. show-models, show-middleware, find-route).