Re-pledging increments the pledge quantity
Resolves #92

This doesn't include any UI changes-that will come in the future
pledging issues. All changes are in the controller and model.

Most of the logic lives in the model. I reorganized Pledge to use
`class << self` because we're amassing a collection of class methods
(and this lets me mark class methods as private).