Refactor controllers
The main goal of this commit is to remove cruft from controllers to make
them easier to read/maintain.

Major changes:

  - Remove unneeded ivars
  - Item can handle its own find_or_create_by_asin logic
  - Set parameter lists in private methods
  * Make wishlist item creation form(s) model-based

This last one might be controversial because this form creates two
objects (potentially)-an Item and a WishlistItem-but only WishlistItems
are given primary status (item fields are disconnected).

I think that a form object is a good idea for the future, but for now,
WishlistItem is the primary goal of this form, and making it a
`form_for` lets us be more succinct in the controller.
`AcceptsNestedAttributesFor` is a mess in terms of error messages, so
I'd like to avoid that.