Standardize Primary-Nav and Clean Up A Little
Our CSS has become a little bit of a mess over the course
of this project. In addition to standardizing link styles
across the site, I'm attempting to bring more clarity to
our style conventions, too. This will be a progress, but
it helps to pay attention to what others have already done
when we add more code.