Improved Desktop Styles
Still not perfection. The navbar stays to the left
as the page expands extra wide, which isn't exactly
what I was going for. Still it's not bad.

I suspect that this quirk will not bother users nearly
as much as it bothers me. This will do just fine for our MVP.