Extract association logic from views
Our partials for pledges and wishlist items had lots of chained
associations. That's brittle (and a little unwieldy), so I've pulled
that logic back into the models, mostly using `delegates`. This way,
we're only talking to the one object in question.

P.S.: There's some display logic tabled between the views and the models
that might be good to extract into a presentation layer at some point.