Add tether for Bootstrap tooltips
We've been getting a console error-Tether is required for Bootstrap
tooltips. Because console errors will fail the (forthcoming) js-enabled
feature specs, it seemed like a good time to do it.

Note: for the future, we should probably prune down the bootstrap
components that we're loading. We're definitely not using all of this.