Users can pledge items
Resolves #93

I'm making a lot of hacky decisions here; I'm not totally comfortable
with this commit. This might need more work before merge.

When a logged-in user clicks on the "Pledge to Donate" button, a pledge
is created, the user's redirected to the pledge page, and another page
is opened up to the Amazon item link.

COMPROMISES/TRADEOFFS:
    - For the sake of time, the JS functionality isn't tested in our
      test suite. I was running into errors with PhantomJS locally, and
      we need to ship something.

    - The user_id (and wishlist_item_id) is a hidden field on the pledge
      button, which means the user could change it if they were feeling
      malicious. I can't think of any real harm that would do, so I'm
      leaving this for now.

    - I haven't done JS with rails for a few years. Am I doing this
      right?

    - Users with JS disabled will not get the new Amazon tab. I'm
      assuming that set is pretty much me, and someone who disables JS is
      perfectly capable of clicking the Amazon link on the pledge page.