Guests can pledge items
Resolves #94

There are still a couple other things I'd like to do before opening the
PR-anyone should be able to unpledge an anonymous pledge, and users
should be able to claim anonymous pledges.

There are some bits that are rough along the edges-the "anonymous
shopping cart" session approach would really be best here-but in the
interest of time, this should be enough to be usable.