Anyone can unpledge an anonymous pledge
If a user wishes to remain anonymous and decides they aren't going to
pledge an item, this will allow them to unpledge the item. Abuse is
possible, but we don't really expect it.

There's an ugly ternary expression in the PledgesController, but I think
it's a short enough method/straightforward enough ternary that it is
still reasonable.

We may want to take a look at the flow through the app in the future.