Simplify ActiveRecordCSVGenerator spec context
The ActiveRecordCSVGenerator spec relied on a lot of doubles in order to
let the tests run without any database interaction. I was also hoping to
free the lib from dependence on either ActiveRecord or this app. The
result was some pretty gnarly code.

Since we're about to pass the baton, it's time to make things easy. If
someone wants to make this test independent from the db/our app in the
future, they can do it in a way that makes sense to them.