Users can claim anonymous pledges
When a user goes to an anonymous pledge page, they can click a "Claim"
button to claim a pledge.

This adds an extra controller action-#claim-to the PledgesController.
I'm not sure how we feel about this approach vs. using a new controller.

A last edge case to address: users claiming pledges when they've already
made a pledge for that item.