Add basic pledge merging
When a user pledges an item, logs out, pledges another copy, logs in,
then tries to claim the pledge, the first pledge is deleted and the
second it claimed with an updated quantity.

This choice was made so that the second pledge url (the one that the
user apparently remembered) is maintained. The downside is that the
"created_at" will be wrong, but hopefully that won't be a big deal.C