Add PR template
We're getting an influx of new people, and it'd be nice to give them a
guide for what we're looking for in our PRs. In particular, the
`Resolves #whatever`, WIP label, and explanation of tradeoffs is super
useful for review.

I decided to put this in .github in order to not clutter up the root.

For more info, see this:
    https://github.com/blog/2111-issue-and-pull-request-templates